implements equals and hashcode

+++ src/main/java/com/golhero/domain/model/MatchId.java
@@ -4,13 +4,13 @@ import java.util.Objects;
 
 public final class MatchId {
 
-    private final int value;
+    private final long value;
 
-    public MatchId(int value) {
+    public MatchId(long value) {
         this.value = value;
     }
 
-    public int getValue() {
+    public long getValue() {
         return value;
     }
